Cleanup and crontool
Contrary to the popular opinion, using setup.py will make everyone's
life easier. Honest. It will.

The new setup.py also creates a console script (available whenever
you're in the virtualenv) called crontool that'll iterate over the
installed scrapers (datasets only for now) and create a pseudo-random
time for it to run.  The output of the command is one cron line per
scraper.  This currently won't be helpful for the following reasons:

1. The run_scraper command doesn't exist yet
2. Even if it did exist the scrapers wouldn't know where to use as their
scratch directory for working in.
3. Absence of deadmansswitch means you'll never know when the cron job
fails to execute (separate from an error caught by run_scraper).