Skeleton (mostly-working) run_scraper script
The new shiny run_scraper script will allow you to run a single scraper,
either all stages or one, and also tell it where it's working directory
is - although currently hard-coded.

Most of the scrapers will now need endpoints and to use the workspace
provided rather than doing it based on code location - but because they
need adding individually it's a good opportunity to check they all
behave the same.